Make version pill fixed to the left side
